Split CuspCorrectionConstruction from LCAOrbitalBuilder.
